:truck: move: テストディレクトリを__test__に変更
ISSUES CLOSED: #130
