+++ src/__test__/components/features/sample/Title.nuxt.spec.ts
@@ -1,7 +1,7 @@
 import { afterEach, beforeEach, describe, expect, test, vi } from 'vitest';
+import { bindTestingPinia, mountSuspendedComponent } from '@/__test__/testHelper';
 import Title from '@/components/features/sample/Title.vue';
 import { useSampleStore } from '@/store/sampleStore';
-import { bindTestingPinia, mountSuspendedComponent } from '@/test/testHelper';
 import type { TestingPinia } from '@pinia/testing';
 
 describe('src/features/sample/components/Title.vue', () => {
